requirements.txt
flask>=2.3.0
requests>=2.28.0
pyyaml>=6.0
pycryptodome>=3.18.0
